spawn de les armes bow i flamethrower arreglat i dany de les armes implementat, spawns random de enemies començat

+++ Assets/Scripts/Enemys/EnemyHealth.cs
@@ -7,7 +7,7 @@ public class EnemyHealth : MonoBehaviour
     [SerializeField] private int startingHealth = 3;
 
     private Flash flash;
-    private int currentHealth;
+    private float currentHealth;
 
     private void Awake()
     {
@@ -20,7 +20,7 @@ public class EnemyHealth : MonoBehaviour
         currentHealth = startingHealth;
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         currentHealth -= damage;
         Debug.Log(currentHealth);
